Arrayify interrupt registers
This commit arrayifies the following interrupt registers:
PROC0_INTE[0-3], PROC0_INTF[0-3], PROC0_INTS[0-3],
PROC1_INTE[0-3], PROC1_INTF[0-3], PROC1_INTS[0-3],
INTR[0-3]
There are still a few more that are are not yet covered.
At the least, the DORMANT_* ones.

The interrupt signals for the RP2040 are spread over several registers.
There are 4 types of interrupt:
  Edge-high, Edge-low, Level-High, Level-Low
and these are sequential bits in the register, presumably so you
can read/write all 4 for a single register in one pass. This does
make it somewhat painful for use to deal with.
To make matters worse, there are only 30 gpio on RP2040, which means
that the last register has more fields than the rest.
In order to get svdtools array to be happy, we have to add fields
for the non-existant GPIO to the last register.

Another option would be to erase the fields entirely, and leave
a 32bit int in it's place. Maybe that's a better option.